refactor (package) : domain package 추가
