+++ src/main/java/ladder/Main.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.domain.Ladder;
+import ladder.domain.User;
 import ladder.view.InputView;
 import ladder.view.OutputView;
