+++ src/main/java/ladder/domain/HorizonLine.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import java.util.ArrayList;
 import java.util.List;
