+++ src/main/java/ladder/domain/ShortLineEnableJudge.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public interface ShortLineEnableJudge {
     boolean isEnable();
