+++ src/test/java/ladder/domain/UserTest.java
@@ -1,5 +1,6 @@
-package ladder;
+package ladder.domain;
 
+import ladder.domain.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
